Refactorice el componente PaymentSection, cree el formulario en un componente FormPayment separado y elimine el código no utilizado
