+++ components/horarios/index.ts
@@ -3,3 +3,4 @@ export { default as DateSelector } from "./DateSelector";
 export { default as TimeSelector } from "./TimeSelector";
 export { default as SeatSelector } from "./SeatSelector";
 export { default as PaymentSection } from "./PaymentSection";
+export { default as FormPayment } from "./FormPayment";
